Move add_unique_parameter_names tests to workflow_spec_utils_test
